Make shared code reusable and handle deployment
Most of the code from the get action can be reused by the other
actions. To do so, move the code to a dedicated common module.

Since wskdeploy can either upload single files or full folders,
every action is made into a folder (package) which includes the
action specific code and the common module.

The deploy script takes care of copying the common module around
in the various action folders. Copies are gitignored so that
we maintain the code in one place only.