Move actions into a single package
Move all different rest actions into a single package so they
can share the common code without the need for it to be copied
around.

The trick to achieve this is to specify a different main function
for each action from the module specified in package.json as the
main module of the package. Functions must be exported to be
used as main functions.